agent_voice_app/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
audioop-lts>=0.2.1; python_version >= "3.13"
